💄 프로필 컴포넌트 분리
